+++ scripts/ex09_velo_optflow.py
@@ -3,8 +3,8 @@
 pyplis example script no. 9 - Plume velocity retrieval using Farneback optical 
 flow algorithm
 """
-from matplotlib.pyplot import close, show
-from os.path import join
+from matplotlib.pyplot import close, show, subplots
+from os.path import join, basename
 import pyplis
 
 ### IMPORT GLOBAL SETTINGS
@@ -12,16 +12,60 @@ from SETTINGS import SAVEFIGS, SAVE_DIR, FORMAT, DPI, OPTPARSE
 
 ### IMPORTS FROM OTHER EXAMPLE SCRIPTS
 from ex04_prep_aa_imglist import prepare_aa_image_list
+SCRIPT_ID = basename(__file__).split("_")[0]
 
 ### SCRIPT OPTONS  
+
+# perform histogram analysis for all images in time series
+HISTO_ANALYSIS_ALL = 1
+HISTO_ANALYSIS_START_IDX = 1
+HISTO_ANALYSIS_STOP_IDX = 207#10
+
+#Gauss pyramid leve
 PYRLEVEL = 1
 BLUR = 0
-IN_ROI = False
-ROI_FLOW = [615, 350, 1230, 790]
+ROI_CONTRAST = [0, 0, 1344, 730]
+MIN_AA = 0.05
+
+PCS1 = pyplis.LineOnImage(345, 350, 450, 195, pyrlevel_def=1, 
+                          line_id="young_plume", color="g")
+PCS2 = pyplis.LineOnImage(80, 10, 80, 270, pyrlevel_def=1, 
+                          line_id="old_plume", color="r")
+    
+LINES = [PCS1, PCS2]
+   
+def analyse_and_plot(lst, lines):
+    mask = lst.get_thresh_mask(MIN_AA)
+    fl=lst.optflow
+    fig, ax = subplots(1,3, figsize=(22,6))
+    fl.plot(ax=ax[0])#, in_roi=True)
+    for line in lines:
+        m = mask * line.get_rotated_roi_mask(fl.flow.shape[:2])
+        line.plot_line_on_grid(ax=ax[0], include_roi_rot=1)
+        #try:
+        _, mu, sigma = fl.plot_orientation_histo(pix_mask=m, 
+                                                 apply_fit=True, ax=ax[1], 
+                                                 color=line.color)
+        low, high = mu-sigma, mu+sigma
+        fl.plot_length_histo(pix_mask=m, apply_fit=False, ax=ax[2], 
+                             dir_low=low, dir_high=high, color=line.color)
+        
+    ax[1].set_title("Orientation histograms")
+    ax[1].set_xlabel(r"$\Theta\,[^{\circ}]$", fontsize=14)
+    ax[1].set_ylabel(r"Count", fontsize=14)
+    ax[2].set_title("Length histograms")
+    ax[2].set_xlabel(r"Magnitude [pix]", fontsize=14)
+    ax[2].set_ylabel(r"Count", fontsize=14)
+    ax[0].get_xaxis().set_ticks([])
+    ax[0].get_yaxis().set_ticks([])
+    fig.tight_layout()
+    return fig
+
     
 ### SCRIPT MAIN FUNCTION
 if __name__ == "__main__":
     close("all")
+    figs = []
     # Prepare aa image list (see example 4)
     aa_list = prepare_aa_image_list()
     
@@ -45,58 +89,75 @@ if __name__ == "__main__":
     # object, meaning, that whenever something changes in "fl", it also does
     # in "aa_list.optflow")
     
-    # Set the region of interest in the optical flow module. The ROI only 
-    # applies to the region used for post analysis of the flow field, the flow
-    # field itself is calculated for the whole image (should become clear from
-    # the plots)
-    if IN_ROI:
-        fl.roi_abs = ROI_FLOW
-    
     # Now activate optical flow calculation in list (this slows down the 
     # speed of the analysis, since the optical flow calculation is 
-    # comparatively slow
-    aa_list.optflow_mode = 1
+    # comparatively slow    
+    s = aa_list.optflow.settings
+    s.hist_dir_gnum_max = 10
+    s.hist_dir_binres = 10
+    s.roi_rad = ROI_CONTRAST
     
-    # Plots the flow field
-    ax0 = fl.draw_flow(0)
+
+    aa_list.optflow_mode = True
     
-    mask = fl.prepare_intensity_condition_mask(lower_val=0.05)
-    count, bins, angles, fit3 = fl.flow_orientation_histo(cond_mask_flat=mask)
-    count, bins, angles, fit4 = fl.flow_length_histo(cond_mask_flat = mask)
+    plume_mask = pyplis.Img(aa_list.get_thresh_mask(MIN_AA))
+    plume_mask.show(tit="AA threshold mask")
     
-    #plot the fit results 
-    ax1 = fit3.plot_result(add_single_gaussians = True)[0]
-    ax2 = fit4.plot_result(add_single_gaussians = True)[0]
+    figs.append(analyse_and_plot(aa_list, LINES))
+        
+    figs.append(fl.plot_flow_histograms(PCS1, plume_mask.img))
+    figs.append(fl.plot_flow_histograms(PCS2, plume_mask.img))
     
     #Show an image containing plume speed magnitudes (ignoring direction)
     velo_img = pyplis.Img(fl.to_plume_speed(dist_img))
-    velo_img.show(vmin = 0, vmax = 7, cmap = "Greens",
+    velo_img.show(vmin=0, vmax = 10, cmap = "Greens",
                   tit = "Optical flow plume velocities",
                   zlabel ="Plume velo [m/s]")
     
-    plume_params = pyplis.plumespeed.LocalPlumeProperties()
-    #mask = fl.prepare_intensity_condition_mask(lower_val = 0.10)
-    plume_params.get_and_append_from_farneback(fl, cond_mask_flat = mask)  
+    # Create two objects used to store time series information about the 
+    # retrieved plume properties
+    plume_props_l1 = pyplis.plumespeed.LocalPlumeProperties(PCS1.line_id)
+    plume_props_l2 = pyplis.plumespeed.LocalPlumeProperties(PCS2.line_id)
     
-    v, verr = plume_params.get_velocity(-1, dist_img.mean())
-    #v = plume_params.len_mu[-1] * dist_img.mean() / fl.del_t
-    fig = fl.plot_flow_histograms()
-    fig.suptitle("v = %.2f (+/- %.2f) m/s" %(v, verr))
+    if HISTO_ANALYSIS_ALL:
+        for k in range(HISTO_ANALYSIS_START_IDX, HISTO_ANALYSIS_STOP_IDX):
+            plume_mask = aa_list.get_thresh_mask(MIN_AA)
+            plume_props_l1.get_and_append_from_farneback(fl, line=PCS1,
+                                                         pix_mask=plume_mask)
+            plume_props_l2.get_and_append_from_farneback(fl, line=PCS2,
+                                                         pix_mask=plume_mask)
+            aa_list.next_img()
+            
+        fig, ax = subplots(2, 1, figsize=(10, 9))
     
-    ### IMPORTANT STUFF FINISHED
-    
+        plume_props_l1.plot_directions(ax=ax[0], 
+                                       color=PCS1.color, 
+                                       label="PCS1")
+        plume_props_l2.plot_directions(ax=ax[0], color=PCS2.color, 
+                                       label="PCS2")
+       
+        plume_props_l1.plot_magnitudes(normalised=True, ax=ax[1], 
+                                      date_fmt="%H:%M:%S", color=PCS1.color, 
+                                      label="PCS1")
+        plume_props_l2.plot_magnitudes(normalised=True, ax=ax[1], 
+                                      date_fmt="%H:%M:%S", color=PCS2.color, 
+                                      label="PCS2") 
+        ax[0].set_xticklabels([])
+        ax[0].legend(loc='best', fancybox=True, framealpha=0.5, fontsize=14) 
+        ax[0].set_title("Movement direction")
+        ax[1].set_title("Displacement length")
+        figs.append(fig)
+        
     if SAVEFIGS:
-        ax0.set_title("")
-        ax0.get_xaxis().set_ticks([])
-        ax0.get_yaxis().set_ticks([])
-        ax0.figure.savefig(join(SAVE_DIR, "ex09_out_1.%s" %FORMAT),
-                           format=FORMAT, dpi=DPI)
-        ax1.figure.savefig(join(SAVE_DIR, "ex09_out_2.%s" %FORMAT),
-                           format=FORMAT, dpi=DPI)
-        ax2.figure.savefig(join(SAVE_DIR, "ex09_out_3.%s" %FORMAT),
-                           format=FORMAT, dpi=DPI)
-        fig.savefig(join(SAVE_DIR, "ex09_out_4.png"))
+        for k in range(len(figs)):
+            figs[k].savefig(join(SAVE_DIR, "%s_out_%d.%s" 
+                            %(SCRIPT_ID, (k+1), FORMAT)),
+                            format=FORMAT, dpi=DPI)
     
+    # Save the time series as txt
+    plume_props_l1.save_txt(SAVE_DIR)
+    plume_props_l2.save_txt(SAVE_DIR)
+
     # Display images or not    
     (options, args)   =  OPTPARSE.parse_args()
     try: