+++ scripts/ex0_6_pcs_lines.py
@@ -26,58 +26,58 @@ def create_example_lines():
     
     cmap = get_cmap("jet")
     # horizontal line, normal orientation to the top (0 deg)
-    lines_r.append(LineOnImage(x0=0, y0=0, x1=1, y1=0, 
+    lines_r.append(LineOnImage(x0=10, y0=10, x1=90, y1=10, 
                                normal_orientation="right",
                                color=cmap(0),
                                line_id="Line 1"))
     
     # horizontal line, normal orientation to the bottom (180 deg)
-    lines_l.append(LineOnImage(0, 0, 1, 0, normal_orientation="left",
+    lines_l.append(LineOnImage(10, 10, 90, 10, normal_orientation="left",
                                color=cmap(0),
                                line_id="Line 1"))
                                
     # Vertical line normal to the right (90 deg)
-    lines_r.append(LineOnImage(0, 0, 0, 1, normal_orientation="right",
+    lines_r.append(LineOnImage(10, 10, 10, 90, normal_orientation="right",
                                color=cmap(50),
                                line_id="Line 2"))
     
     # Vertical line normal to the left (270 deg)
-    lines_l.append(LineOnImage(0, 0, 0, 1, normal_orientation="left",
+    lines_l.append(LineOnImage(10, 10, 10, 90, normal_orientation="left",
                                color=cmap(50),
                                line_id="Line 2"))
     
     # Slanted line 45 degrees
-    lines_r.append(LineOnImage(.1, .3, .6, .8, normal_orientation="right",
+    lines_r.append(LineOnImage(20, 30, 50, 60, normal_orientation="right",
                                color=cmap(100),
                                line_id="Line 3"))
     
     # Slanted line 45 degrees
-    lines_l.append(LineOnImage(.1, .3, .6, .8, normal_orientation="left",
+    lines_l.append(LineOnImage(20, 30, 50, 60, normal_orientation="left",
                                color=cmap(100),
                                line_id="Line 3"))
     
     # Slanted line 45 degrees
-    lines_r.append(LineOnImage(1, 0, 0, 1, normal_orientation="right",
+    lines_r.append(LineOnImage(90, 10, 10, 90, normal_orientation="right",
                                color=cmap(150),
                                line_id="Line 4"))
     
     # Slanted line 45 degrees
-    lines_l.append(LineOnImage(1, 0, 0, 1, normal_orientation="left",
+    lines_l.append(LineOnImage(90, 10, 10, 90, normal_orientation="left",
                                color=cmap(150),
                                line_id="Line 4"))
     
-    lines_r.append(LineOnImage(0.6, .1, .8, .9, normal_orientation="right",
+    lines_r.append(LineOnImage(60, 20, 80, 90, normal_orientation="right",
                                color=cmap(200),
                                line_id="Line 5"))
                                
-    lines_l.append(LineOnImage(0.6, .1, .8, .9, normal_orientation="left",
+    lines_l.append(LineOnImage(60, 20, 80, 90, normal_orientation="left",
                                color=cmap(200),
                                line_id="Line 5"))
     
-    lines_r.append(LineOnImage(0.4, .1, .2, .9, normal_orientation="right",
+    lines_r.append(LineOnImage(40, 20, 30, 90, normal_orientation="right",
                                color=cmap(250),
                                line_id="Line 6"))
-    lines_l.append(LineOnImage(0.4, .1, .2, .9, normal_orientation="left",
+    lines_l.append(LineOnImage(40, 20, 30, 90, normal_orientation="left",
                                color=cmap(250),
                                line_id="Line 6"))
     
@@ -93,22 +93,24 @@ if __name__ == "__main__":
         #print "%d: %s" %(k, line.orientation_info)
         normal = line.normal_vector
         lbl = "%s" %line.line_id
-        line.plot_line_on_grid(ax = ax[0], include_normal = 1, label = lbl)
+        line.plot_line_on_grid(ax=ax[0], include_normal=1, 
+                               include_roi_rot=True, label=lbl)
     for k in range(len(lines_l)):
         line = lines_l[k]
         normal = line.normal_vector
         lbl = "%s" %line.line_id
-        line.plot_line_on_grid(ax = ax[1], include_normal = 1, label = lbl)
+        line.plot_line_on_grid(ax=ax[1], include_normal=1, 
+                               include_roi_rot=True, label=lbl)
     
     ax[0].set_title("Orientation right")
     ax[0].legend(loc = "best", fontsize=8, framealpha = 0.5)
-    ax[0].set_xlim([-.1, 1.1])
-    ax[0].set_ylim([1.1, -.1])
+    ax[0].set_xlim([0, 100])
+    ax[0].set_ylim([100, 0])
     
     ax[1].set_title("Orientation left")
     ax[1].legend(loc = "best", fontsize=8, framealpha = 0.5)
-    ax[1].set_xlim([-.1, 1.1])
-    ax[1].set_ylim([1.1, -.1])
+    ax[1].set_xlim([0, 100])
+    ax[1].set_ylim([100, 0])
     ### IMPORTANT STUFF FINISHED    
     if SAVEFIGS:
         fig.savefig(join(SAVE_DIR, "ex0_6_out_1.%s" %FORMAT),